+++ d2.config.js
@@ -4,7 +4,9 @@ const config = {
     id: 'a4cd3827-e717-4e09-965d-ab05df2591e5',
     title: 'Line Listing',
 
-    minDHIS2Version: '2.39',
+    minDHIS2Version: '2.40',
+
+    pluginType: 'DASHBOARD',
 
     pwa: {
         enabled: true,